Update .gitignore; added error handling to output

+++ utils/file.py
@@ -2,21 +2,33 @@ import os
 from utils.error import error
 
 def create_file(file: str) -> int:
-    file = open(f"{file}", "w+")
-    file.close()
-    return 0
+    if os.path.exists(file):
+        file = open(f"{file}", "w+")
+        file.close()
+        return 0
+    else:
+        error(file, 1, "file")
+        return 1
 
 def write_file(file: str, text: str) -> int:
-    file = open(f"{file}", "w")
-    file.write(text)
-    file.close()
-    return 0
+    if os.path.exists(file):
+        file = open(f"{file}", "w")
+        file.write(text)
+        file.close()
+        return 0
+    else:
+        error(file, 1, "file")
+        return 1
 
 def read_file(file: str) -> int:
-    file = open(f"{file}", "r")
-    print(file.read())
-    file.close()
-    return 0
+    if os.path.exists(file):
+        file = open(f"{file}", "r")
+        print(file.read())
+        file.close()
+        return 0
+    else:
+        error(file, 1, "file")
+        return 1
 
 def remove_file(file: str) -> int:
     if os.path.exists(file):
